fix: resolve Linux cross-compile warnings in our code
Fixes for clang-cl-only diagnostics surfaced by the Linux
cross-compile job (rarely triggered, since it only runs on PRs
touching vcpkg.json/CMake/extern paths):

- DX12SwapChain.h: 9 IDXGISwapChain overrides were missing the
  override keyword (-Winconsistent-missing-override).
- ShaderCache.h, PerformanceOverlay.h, ABTesting.h: LARGE_INTEGER
  fields brace-initialized as `= { 0 }` only value-initialize the
  union's first alternative, leaving HighPart indeterminate
  (-Wmissing-field-initializers); switched to `{}`.
- EditorWindow.cpp, WeatherWidget.cpp: loop counters compared
  against RE::TESWeather::ColorTimes::ColorTime (uint32_t enum)
  as a plain int (-Wsign-compare). Sites that only index an array
  are now size_t; the two that feed an int-parameter helper keep
  int with an explicit cast at the comparison.
- OpenDRT.cpp: all 70 ImGui::Text(T(...)) calls pass a runtime
  (translated) string as the format argument with no format args
  of their own (-Wformat-security) -- a latent format-string risk,
  not just a warning. Switched to ImGui::TextUnformatted, which
  doesn't interpret its argument as a format string at all. Also
  removed 3 unused const label arrays (-Wunused-const-variable),
  two referenced only from an already-commented-out block.

Verified: every changed file rebuilds warning-free under MSVC
/W4 /WX (the Linux job's clang-cl-specific diagnostics can't be
reproduced locally; CI is the real verification for those).

+++ src/CSEditor/EditorWindow.cpp
@@ -508,7 +508,7 @@ void EditorWindow::ShowObjectsWindow()
 				};
 
 				auto addTOD = [&](auto*(&fields)[RE::TESWeather::ColorTimes::kTotal], const WidgetVec& widgets) {
-					for (int tod = 0; tod < RE::TESWeather::ColorTimes::kTotal; ++tod) {
+					for (int tod = 0; tod < static_cast<int>(RE::TESWeather::ColorTimes::kTotal); ++tod) {
 						auto* form = fields[tod];
 						if (!form)
 							continue;
